[refactor] 당첨 통계를 관리하는 Model(WinTicket) 분리
- 기존에 Controller에 존재하던 당첨 통계를 계산하고,
  관련된 데이터를 관리하는 로직을 분리함
- 이로써 Controller에는 View와 Model을 다루는 로직만 남아있게 됨
